refactor(imports): reorganize and group imports and add missing event_model.py
